- Removed service interfaces, as suggested by Jan Ka on Facebook - Moved OpenAPI and security related classes to more descriptive packages - Created custom OpenAPI annotations for easier documentation of controller classes - Added Hibernate metamodel generator dependency for easier creation of Specification objects
