+++ src/main/java/com/eukon05/financetracker/openapi/param/FromDateParam.java
@@ -1,0 +1,16 @@
+package com.eukon05.financetracker.openapi.param;
+
+import io.swagger.v3.oas.annotations.Parameter;
+import io.swagger.v3.oas.annotations.media.Schema;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+import java.time.Instant;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.METHOD)
+@Parameter(name = "fromDate", schema = @Schema(implementation = Instant.class))
+public @interface FromDateParam {
+}
