+++ src/main/java/com/eukon05/financetracker/openapi/param/HideParamsMap.java
@@ -1,0 +1,14 @@
+package com.eukon05.financetracker.openapi.param;
+
+import io.swagger.v3.oas.annotations.Parameter;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.METHOD)
+@Parameter(name = "params", hidden = true)
+public @interface HideParamsMap {
+}
