+++ src/main/java/com/eukon05/financetracker/openapi/param/ValueParam.java
@@ -1,0 +1,16 @@
+package com.eukon05.financetracker.openapi.param;
+
+import io.swagger.v3.oas.annotations.Parameter;
+import io.swagger.v3.oas.annotations.media.Schema;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+import java.math.BigDecimal;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.METHOD)
+@Parameter(name = "value", schema = @Schema(implementation = BigDecimal.class))
+public @interface ValueParam {
+}
