+++ src/main/java/com/eukon05/financetracker/openapi/response/AccessDeniedResponse.java
@@ -1,0 +1,15 @@
+package com.eukon05.financetracker.openapi.response;
+
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.METHOD)
+@ApiResponse(responseCode = "403", description = "Access denied", content = @Content)
+public @interface AccessDeniedResponse {
+}
