+++ src/main/java/com/eukon05/financetracker/openapi/response/NotFoundResponse.java
@@ -1,0 +1,17 @@
+package com.eukon05.financetracker.openapi.response;
+
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import org.springframework.http.ProblemDetail;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.METHOD)
+@ApiResponse(responseCode = "404", description = "Entity not found or does not belong to your account", content = @Content(schema = @Schema(implementation = ProblemDetail.class)))
+public @interface NotFoundResponse {
+}
