+++ src/main/java/com/eukon05/financetracker/security/CorsConfiguration.java
@@ -1,4 +1,4 @@
-package com.eukon05.financetracker.configuration;
+package com.eukon05.financetracker.security;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -9,7 +9,7 @@ import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 class CorsConfiguration {
 
     @Bean
-    public WebMvcConfigurer corsConfigurer() {
+    WebMvcConfigurer corsConfigurer() {
         return new WebMvcConfigurer() {
             @Override
             public void addCorsMappings(CorsRegistry registry) {
