+++ src/main/java/com/eukon05/financetracker/transaction/TransactionController.java
@@ -1,11 +1,11 @@
 package com.eukon05.financetracker.transaction;
 
+import com.eukon05.financetracker.openapi.response.NotFoundResponse;
+import com.eukon05.financetracker.openapi.response.ValidationErrorResponse;
 import com.eukon05.financetracker.transaction.dto.CreateTransactionDTO;
 import com.eukon05.financetracker.transaction.dto.EditTransactionDTO;
 import com.eukon05.financetracker.transaction.dto.TransactionDTO;
 import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -26,11 +26,8 @@ class TransactionController {
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
     @Operation(summary = "Create a new transaction")
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "400", ref = "validation"),
-            @ApiResponse(responseCode = "401", ref = "unauthorized"),
-            @ApiResponse(responseCode = "404", ref = "notfound")
-    })
+    @ValidationErrorResponse
+    @NotFoundResponse
     void createTransaction(@RequestBody @Valid CreateTransactionDTO dto, @AuthenticationPrincipal Jwt jwt) {
         facade.createTransaction(jwt.getSubject(), dto);
     }
@@ -38,10 +35,7 @@ class TransactionController {
     @DeleteMapping("/{transactionID}")
     @ResponseStatus(HttpStatus.OK)
     @Operation(summary = "Delete transaction with a given ID")
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "401", ref = "unauthorized"),
-            @ApiResponse(responseCode = "404", ref = "notfound")
-    })
+    @NotFoundResponse
     void deleteTransaction(@PathVariable long transactionID, @AuthenticationPrincipal Jwt jwt) {
         facade.deleteTransaction(jwt.getSubject(), transactionID);
     }
@@ -49,10 +43,7 @@ class TransactionController {
     @GetMapping(value = "/{transactionID}", produces = MediaType.APPLICATION_JSON_VALUE)
     @ResponseStatus(HttpStatus.OK)
     @Operation(summary = "Get information about a transaction with a given ID")
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "401", ref = "unauthorized"),
-            @ApiResponse(responseCode = "404", ref = "notfound")
-    })
+    @NotFoundResponse
     TransactionDTO getTransaction(@PathVariable long transactionID, @AuthenticationPrincipal Jwt jwt) {
         return facade.getTransactionDTO(jwt.getSubject(), transactionID);
     }
@@ -60,11 +51,8 @@ class TransactionController {
     @PatchMapping("/{transactionID}")
     @ResponseStatus(HttpStatus.OK)
     @Operation(summary = "Edit transaction with a given ID")
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "400", ref = "validation"),
-            @ApiResponse(responseCode = "401", ref = "unauthorized"),
-            @ApiResponse(responseCode = "404", ref = "notfound")
-    })
+    @ValidationErrorResponse
+    @NotFoundResponse
     void editTransaction(@PathVariable long transactionID, @AuthenticationPrincipal Jwt jwt, @RequestBody @Valid EditTransactionDTO dto) {
         facade.editTransaction(jwt.getSubject(), transactionID, dto);
     }
