+++ src/main/java/com/eukon05/financetracker/transaction/TransactionFacade.java
@@ -3,22 +3,85 @@ package com.eukon05.financetracker.transaction;
 import com.eukon05.financetracker.transaction.dto.CreateTransactionDTO;
 import com.eukon05.financetracker.transaction.dto.EditTransactionDTO;
 import com.eukon05.financetracker.transaction.dto.TransactionDTO;
+import com.eukon05.financetracker.transaction_category.TransactionCategory;
+import com.eukon05.financetracker.transaction_category.TransactionCategoryFacade;
+import com.eukon05.financetracker.transaction_category.exception.TransactionCategoryNotFoundException;
+import com.eukon05.financetracker.wallet.Wallet;
+import com.eukon05.financetracker.wallet.WalletFacade;
+import com.eukon05.financetracker.wallet.exception.WalletNotFoundException;
+import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Map;
 
-public interface TransactionFacade {
-    void createTransaction(String userId, CreateTransactionDTO dto);
+@Service
+@RequiredArgsConstructor
+public class TransactionFacade {
+    private final TransactionService transactionService;
+    private final WalletFacade walletFacade;
+    private final TransactionCategoryFacade categoryFacade;
 
-    void deleteTransaction(String userId, long transactionID);
+    private Transaction getTransaction(String userId, long transactionId) {
+        return transactionService.getTransaction(userId, transactionId);
+    }
 
-    void editTransaction(String userId, long transactionID, EditTransactionDTO dto);
+    @Transactional
+    public void createTransaction(String userId, CreateTransactionDTO dto) {
+        Wallet wallet = walletFacade.getWallet(userId, dto.walletID());
+        TransactionCategory category = categoryFacade.getTransactionCategory(dto.categoryID());
 
-    TransactionDTO getTransactionDTO(String userId, long transactionID);
+        transactionService.createTransaction(wallet, category, dto);
+    }
 
-    Page<TransactionDTO> getPagedTransactionDTOsForWallet(String userId, long walletID, Map<String, String> params, Pageable pageable);
+    public void deleteTransaction(String userId, long transactionID) {
+        Transaction transaction = getTransaction(userId, transactionID);
+        transactionService.deleteTransaction(transaction);
+    }
 
-    List<Transaction> getTransactionsForSpecification(String userId, long walletID, Map<String, String> params);
+    @Transactional
+    public void editTransaction(String userId, long transactionID, EditTransactionDTO dto) {
+        Transaction transaction = getTransaction(userId, transactionID);
+        TransactionCategory category = categoryFacade.getTransactionCategory(dto.categoryID());
+        transactionService.editTransaction(transaction, category, dto);
+    }
+
+    public TransactionDTO getTransactionDTO(String userId, long transactionID) {
+        return transactionService.getTransactionDTO(userId, transactionID);
+    }
+
+    public Page<TransactionDTO> getPagedTransactionDTOsForWallet(String userId, long walletID, Map<String, String> params, Pageable pageable) {
+        checkUserOwnsWallet(userId, walletID);
+
+        if (params.containsKey(Transaction_.CATEGORY)) {
+            checkCategoryExists(Long.parseLong(params.get(Transaction_.CATEGORY)));
+        }
+
+        return transactionService.getPagedTransactionDTOsForSpecification(TransactionSpecification.build(walletID, params), pageable);
+    }
+
+    public List<Transaction> getTransactionsForSpecification(String userId, long walletID, Map<String, String> params) {
+        checkUserOwnsWallet(userId, walletID);
+
+        if (params.containsKey(Transaction_.CATEGORY)) {
+            checkCategoryExists(Long.parseLong(params.get(Transaction_.CATEGORY)));
+        }
+
+        return transactionService.getTransactionsForSpecification(TransactionSpecification.build(walletID, params));
+    }
+
+    private void checkUserOwnsWallet(String userId, long walletID) {
+        if (!walletFacade.checkOwnership(userId, walletID)) {
+            throw new WalletNotFoundException(walletID);
+        }
+    }
+
+    private void checkCategoryExists(long categoryID) {
+        if (!categoryFacade.exists(categoryID)) {
+            throw new TransactionCategoryNotFoundException(categoryID);
+        }
+    }
 }
