+++ src/main/java/com/eukon05/financetracker/transaction_category/TransactionCategoryController.java
@@ -1,11 +1,13 @@
 package com.eukon05.financetracker.transaction_category;
 
+import com.eukon05.financetracker.openapi.response.AccessDeniedResponse;
+import com.eukon05.financetracker.openapi.response.ConflictResponse;
+import com.eukon05.financetracker.openapi.response.NotFoundResponse;
+import com.eukon05.financetracker.openapi.response.ValidationErrorResponse;
 import com.eukon05.financetracker.transaction_category.dto.CreateTransactionCategoryDTO;
 import com.eukon05.financetracker.transaction_category.dto.EditTransactionCategoryDTO;
 import com.eukon05.financetracker.transaction_category.dto.TransactionCategoryDTO;
 import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -24,14 +26,9 @@ class TransactionCategoryController {
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
     @Operation(summary = "Create a new transaction category")
-    @ApiResponses(
-            value = {
-                    @ApiResponse(responseCode = "400", ref = "validation"),
-                    @ApiResponse(responseCode = "401", ref = "unauthorized"),
-                    @ApiResponse(responseCode = "403", ref = "unauthorized"),
-                    @ApiResponse(responseCode = "409", ref = "conflict")
-            }
-    )
+    @ValidationErrorResponse
+    @ConflictResponse
+    @AccessDeniedResponse
     void createTransactionCategory(@RequestBody @Valid CreateTransactionCategoryDTO dto) {
         facade.createTransactionCategory(dto);
     }
@@ -39,12 +36,7 @@ class TransactionCategoryController {
     @GetMapping
     @ResponseStatus(HttpStatus.OK)
     @Operation(summary = "Get information about all transaction categories")
-    @ApiResponses(
-            value = {
-                    @ApiResponse(responseCode = "401", ref = "unauthorized"),
-                    @ApiResponse(responseCode = "403", ref = "unauthorized")
-            }
-    )
+    @AccessDeniedResponse
     List<TransactionCategoryDTO> getAllCategories(@RequestParam(name = "type", required = false) TransactionCategoryType type) {
         return facade.getTransactionCategoryDTOs(type);
     }
@@ -52,13 +44,8 @@ class TransactionCategoryController {
     @GetMapping("/{id}")
     @ResponseStatus(HttpStatus.OK)
     @Operation(summary = "Get information about a single transaction category with a given ID")
-    @ApiResponses(
-            value = {
-                    @ApiResponse(responseCode = "404", ref = "notfound"),
-                    @ApiResponse(responseCode = "401", ref = "unauthorized"),
-                    @ApiResponse(responseCode = "403", ref = "unauthorized"),
-            }
-    )
+    @NotFoundResponse
+    @AccessDeniedResponse
     TransactionCategoryDTO getCategory(@PathVariable long id) {
         return facade.getTransactionCategoryDTO(id);
     }
@@ -66,13 +53,8 @@ class TransactionCategoryController {
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.OK)
     @Operation(summary = "Delete transaction category with a given ID")
-    @ApiResponses(
-            value = {
-                    @ApiResponse(responseCode = "404", ref = "notfound"),
-                    @ApiResponse(responseCode = "401", ref = "unauthorized"),
-                    @ApiResponse(responseCode = "403", ref = "unauthorized"),
-            }
-    )
+    @NotFoundResponse
+    @AccessDeniedResponse
     void deleteTransactionCategory(@PathVariable long id) {
         facade.deleteTransactionCategory(id);
     }
@@ -80,15 +62,10 @@ class TransactionCategoryController {
     @PatchMapping("/{id}")
     @ResponseStatus(HttpStatus.OK)
     @Operation(summary = "Edit transaction category with a given ID")
-    @ApiResponses(
-            value = {
-                    @ApiResponse(responseCode = "404", ref = "notfound"),
-                    @ApiResponse(responseCode = "401", ref = "unauthorized"),
-                    @ApiResponse(responseCode = "403", ref = "unauthorized"),
-                    @ApiResponse(responseCode = "400", ref = "validation"),
-                    @ApiResponse(responseCode = "409", ref = "conflict")
-            }
-    )
+    @NotFoundResponse
+    @AccessDeniedResponse
+    @ValidationErrorResponse
+    @ConflictResponse
     void editTransactionCategory(@PathVariable long id, @RequestBody @Valid EditTransactionCategoryDTO dto) {
         facade.editTransactionCategory(id, dto);
     }
